Fix standalone meson build (#308)
Add fix for standalone build for mctpd repo, as changes done in
I3CDriver.cpp as the type is changed from int to long int.

TESTED :
1. On build server meson build is successful.

+++ src/hw/aspeed/I3CDriver.cpp
@@ -184,7 +184,7 @@ void I3CDriver::discoverI3CDevices()
                     "Error registering MCTP o. I3C default client");
                 return;
             }
-            streamMonitor.assign(reinterpret_cast<int>(handle.release()));
+            streamMonitor.assign(reinterpret_cast<long int>(handle.release()));
             break;
         }
         else
